Migrate pathing to it's own library
